+++ CHANGELOG.md
@@ -13,7 +13,7 @@
 
 ## 1.3.0 (2023-12-05)
 
-* added support for form recovery. Upon reconnection, the client sends an event (`options_recovery`) that contains the latest selection. This allows recovery of the options that were active before the view disconnected.
+* added support for selection recovery. Upon reconnection, the client sends an event (`selection_recovery`) that contains the latest selection. This allows recovery of the selection that was active before the view disconnected.
 
 ## 1.2.2 (2023-10-21)